+++ frontend/src/components/ReviewListBack.vue
@@ -1,28 +1,64 @@
 <template>
+  <form style="color: white; width: 90%; height: 180px; margin: 50px auto auto auto">
+    <div style="display: flex">
+      <div class="mb-3">
+        <input type="email" placeholder="리뷰 제목" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp">
+      </div>
+      <div style="margin-bottom: 10px">
+        <v-rating
+            v-model="rating"
+            class="ma-2"
+            density="comfortable"
+        ></v-rating>
+      </div>
+
+    </div>
+    <div class="mb-3">
+      <input type="password" placeholder="리뷰 내용" class="form-control" id="exampleInputPassword1">
+    </div>
+    <div style="text-align: right;">
+      <button type="submit" class="btn btn-primary"
+              style="background-color: #392365; border-color: #392365; margin-left: 15px;">리뷰 남기기
+      </button>
+    </div>
+
+  </form>
   <div class="review-list">
     <div class="review-grid">
+      <div>
+
+      </div>
       <v-row>
         <v-col v-for="review in reviewList" :key="review.id" cols="12">
-          <v-card class="review-card">
-            <v-card-title>{{ review.title }}</v-card-title>
-            <v-card-subtitle>별점 : {{ displayRating(review.rating) }}</v-card-subtitle>
-            <v-card-text>수정일 : {{ review.modifiedAt }}</v-card-text>
-            <v-card-text>닉네임 : {{ review.nickname }}</v-card-text>
-            <v-card-text>내용 : {{ review.content }}</v-card-text>
+          <v-card class="review-card" style="width: 1050px; height: 150px">
+            <div style="display: flex;">
+              <v-card-title>{{ review.title }}</v-card-title>
+              <div style="margin-top: 13px">
+                <v-card-subtitle>{{ displayRating(review.rating) }}</v-card-subtitle>
+              </div>
+              <v-card-text>{{ review.nickname }}</v-card-text>
+              <div style="margin-left: 50px">
+                <v-card-text>{{ review.modifiedAt }}</v-card-text>
+              </div>
+            </div>
+            <v-card-text>{{ review.content }}</v-card-text>
           </v-card>
         </v-col>
       </v-row>
     </div>
-<!--    &lt;!&ndash; 페이지 네이션 &ndash;&gt;-->
-<!--    <div class="pagination">-->
-<!--      <v-btn @click="fetchReviews(currentPage - 1)" :disabled="currentPage === 0">-->
-<!--        이전 페이지-->
-<!--      </v-btn>-->
-<!--      <span>현재 페이지: {{ currentPage + 1 }}</span>-->
-<!--      <v-btn @click="fetchReviews(currentPage + 1)" :disabled="currentPage === totalPages - 1">-->
-<!--        다음 페이지-->
-<!--      </v-btn>-->
-<!--    </div>-->
+    <!-- 페이지 네이션 -->
+    <div style="width: 100%; margin: 10px">
+      <div class="pagination" style="margin-left: 447px ">
+        <v-btn @click="fetchReviews(currentPage - 1)" :disabled="currentPage === 0"
+               style="background-color: #0a0925; color: white;">
+          이전
+        </v-btn>
+        <v-btn @click="fetchReviews(currentPage + 1)" :disabled="currentPage === totalPages - 1"
+               style="margin-left: 20px; background-color: #0a0925; color: white;">
+          다음
+        </v-btn>
+      </div>
+    </div>
   </div>
 
   <div class="text-center">
@@ -32,7 +68,7 @@
           <v-container class="max-width">
             <v-pagination
                 v-model="page"
-                :length="20"
+                :length="totalPages"
                 class="my-4"
             ></v-pagination>
           </v-container>
@@ -44,11 +80,18 @@
 </template>
 
 <script>
-import { ref, onMounted } from 'vue';
+import {ref, onMounted} from 'vue';
 import axios from 'axios';
-import { useRouter } from 'vue-router';
+import {useRouter} from 'vue-router';
 
 export default {
+  data() {
+    return {
+      rating: 3,
+      page: 1,
+    }
+  },
+
   setup() {
     const reviewList = ref([]);
     const currentPage = ref(0);
@@ -81,7 +124,7 @@ export default {
       fetchReviews(currentPage.value);
     });
 
-    return { reviewList, currentPage, totalPages, fetchReviews, displayRating };
+    return {reviewList, currentPage, totalPages, fetchReviews, displayRating};
   }
 }
 </script>
