+++ frontend/src/css/styles/TestStyle.css
@@ -1,6 +1,6 @@
 /* 기본 글로벌 스타일 */
 main {
-  height: 100%;
+  height: 2300px;
   background-color: #c6c9e7;
 }
