--- frontend/src/views/HomeViewBack.vue
@@ -1,11 +1,0 @@
-<script setup>
-
-</script>
-
-<template>
-
-</template>
-
-<style scoped>
-
-</style>
