+++ frontend/src/views/HomeViewFront.vue
@@ -9,7 +9,8 @@
   </head>
   <body>
   <main>
-    <div style="height: 1300px; width: 100%; padding: 100px;">
+    <div class="top5">
+      <h1>평점 순 TOP.5</h1>
       <div style="width: 1600px; padding:10px; display: flex; margin: 0 auto;">
         <CardView/>
         <CardView/>
@@ -17,6 +18,19 @@
         <CardView/>
         <CardView/>
       </div>
+    </div>
+    <div class="top5">
+      <h1>좋아요 순 TOP.5</h1>
+      <div style="width: 1600px; padding:10px; display: flex; margin: 0 auto;">
+        <CardView/>
+        <CardView/>
+        <CardView/>
+        <CardView/>
+        <CardView/>
+      </div>
+    </div>
+    <div class="top5">
+      <h1>오픈예정일 순 TOP.5</h1>
       <div style="width: 1600px; padding:10px; display: flex; margin: 0 auto;">
         <CardView/>
         <CardView/>
@@ -35,9 +49,16 @@
 </template>
 <style>
 @import "../css/styles/TestStyle.css";
+
+.top5 {
+  height: 700px;
+  width: 100%;
+  padding: 100px;
+}
 </style>
 <script>
 import CardView from "@/components/EventCardFront.vue";
+
 export default {
   name: 'App',
