+++ src/main/kotlin/com/codersgate/ticketraider/domain/event/model/Event.kt
@@ -47,7 +47,7 @@ class Event(
     @JoinColumn(name = "category_id")
     var category: Category,
 
-    @OneToOne(mappedBy = "event", fetch = FetchType.LAZY, cascade = [CascadeType.ALL], orphanRemoval = true)
+    @OneToOne(mappedBy = "event", fetch = FetchType.EAGER, cascade = [CascadeType.ALL], orphanRemoval = true)
     var price: Price? = null,
 
     @OneToMany(mappedBy = "event", fetch = FetchType.EAGER, cascade = [CascadeType.ALL], orphanRemoval = true)
