+++ src/main/kotlin/com/codersgate/ticketraider/domain/event/service/EventService.kt
@@ -2,6 +2,7 @@ package com.codersgate.ticketraider.domain.event.service
 
 import com.codersgate.ticketraider.domain.event.dto.EventRequest
 import com.codersgate.ticketraider.domain.event.dto.EventResponse
+import com.codersgate.ticketraider.domain.event.dto.price.PriceResponse
 import org.springframework.data.domain.Page
 import org.springframework.stereotype.Service
 import org.springframework.data.domain.Pageable
@@ -9,6 +10,8 @@ import org.springframework.web.multipart.MultipartFile
 
 @Service
 interface EventService {
+
+    fun getPrice(eventId: Long) : PriceResponse
     fun createEvent(eventRequest: EventRequest, file: MultipartFile?)
 
     fun updateEvent(eventId: Long, eventRequest: EventRequest, file: MultipartFile?)
